requirements.txt
altgraph

autopep8

contourpy

cycler

filetype

fonttools

kiwisolver

matplotlib

numpy

opencv-python

packaging

pandas

pefile

Pillow

pycodestyle

pyinstaller

pyinstaller-hooks-contrib

pyparsing

pytesseract

python-dateutil

pytz

pywin32-ctypes

six

tzdata


